Add `use_debanding` render mode and property to spatial shaders
This works with all rendering methods, but is most suited to the Mobile
and Compatibility rendering methods which lack precision in their color buffer.

When using Forward+, this can still be useful to prevent occasional ringing
artifacts that occur with viewport debanding (at the cost of a slightly more
visible dithering pattern).

It is possible to use both material and viewport-based debanding at the
same time, but it's not recommended as it doesn't bring much of a benefit.